move flags class to its own package
