+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/AccountDelete.java
@@ -6,7 +6,8 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.primitives.UnsignedInteger;
 import org.immutables.value.Value;
 import org.immutables.value.Value.Derived;
-import org.xrpl.xrpl4j.model.transactions.Flags.TransactionFlags;
+import org.xrpl.xrpl4j.model.flags.Flags;
+import org.xrpl.xrpl4j.model.flags.Flags.TransactionFlags;
 
 import java.util.Optional;
