+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/DepositPreAuth.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.base.Preconditions;
 import org.immutables.value.Value;
 import org.immutables.value.Value.Derived;
+import org.xrpl.xrpl4j.model.flags.Flags;
 
 import java.util.Optional;
