+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/EscrowCancel.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.primitives.UnsignedInteger;
 import org.immutables.value.Value;
+import org.xrpl.xrpl4j.model.flags.Flags;
 
 /**
  * Return escrowed XRP to the sender.
