+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/OfferCancel.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.primitives.UnsignedInteger;
 import org.immutables.value.Value;
 import org.immutables.value.Value.Immutable;
+import org.xrpl.xrpl4j.model.flags.Flags;
 
 import java.util.Optional;
