+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/PaymentChannelClaim.java
@@ -4,7 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import org.immutables.value.Value;
-import org.xrpl.xrpl4j.model.transactions.Flags.PaymentChannelClaimFlags;
+import org.xrpl.xrpl4j.model.flags.Flags;
+import org.xrpl.xrpl4j.model.flags.Flags.PaymentChannelClaimFlags;
 
 import java.util.Optional;
