+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/PaymentChannelCreate.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.primitives.UnsignedInteger;
 import com.google.common.primitives.UnsignedLong;
 import org.immutables.value.Value;
+import org.xrpl.xrpl4j.model.flags.Flags;
 
 import java.util.Optional;
