+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/transactions/SetRegularKey.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import org.immutables.value.Value;
+import org.xrpl.xrpl4j.model.flags.Flags;
 
 import java.util.Optional;
