+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/flags/SignerListFlagsTests.java
@@ -1,10 +1,12 @@
-package org.xrpl.xrpl4j.model.transactions;
+package org.xrpl.xrpl4j.model.flags;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
+import org.xrpl.xrpl4j.model.flags.Flags;
+import org.xrpl.xrpl4j.model.transactions.AbstractFlagsTest;
 
 import java.util.Collection;
